--- a/sm2/sm2p256_amd64.go
+++ b/sm2/sm2p256_amd64.go
@@ -35,6 +35,7 @@ var (
 	p256            p256Curve
 	p256Precomputed *[37][64 * 8]uint64
 	precomputeOnce  sync.Once
+	SM2PARAM_A      []byte
 )
 
 func initP256() {
@@ -46,6 +47,9 @@ func initP256() {
 	p256.Gx, _ = new(big.Int).SetString("32C4AE2C1F1981195F9904466A39C9948FE30BBFF2660BE1715A4589334C74C7", 16)
 	p256.Gy, _ = new(big.Int).SetString("BC3736A2F4F6779C59BDCEE36B692153D0A9877CC62A474002DF32E52139F0A0", 16)
 	p256.BitSize = 256
+	A, _ := new(big.Int).SetString("FFFFFFFEFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF00000000FFFFFFFFFFFFFFFC", 16)
+	SM2PARAM_A = A.Bytes()
+	precomputeOnce.Do(initTable)
 }
 
 func (curve p256Curve) Params() *elliptic.CurveParams {
@@ -262,7 +266,7 @@ func maybeReduceModP(in *big.Int) *big.Int {
 //	sum.CopyConditional(&r2, r1IsInfinity)
 //	return sum.p256PointToAffine()
 //}
-func (curve p256Curve) CombinedMult(Precomputed *[37][64*8]uint64, baseScalar, scalar []byte) (x, y *big.Int) {
+func (curve p256Curve) CombinedMult(Precomputed *[37][64 * 8]uint64, baseScalar, scalar []byte) (x, y *big.Int) {
 	scalarReversed := make([]uint64, 4)
 	var r1 p256Point
 	r2 := new(p256Point)
@@ -291,7 +295,7 @@ func (curve p256Curve) CombinedMult(Precomputed *[37][64*8]uint64, baseScalar, s
 	////sm2p256PointAddAsm(r1.xyz[:], r1.xyz[:], r2.xyz[:])
 
 	//r2.p256ScalarMult(scalarReversed)
-	r2.p256PreMult(Precomputed,scalarReversed)
+	r2.p256PreMult(Precomputed, scalarReversed)
 
 	var sum, double p256Point
 	pointsEqual := sm2p256PointAddAsm(sum.xyz[:], r1.xyz[:], r2.xyz[:])
@@ -639,8 +643,6 @@ func initTable() {
 }
 
 func (p *p256Point) p256BaseMult(scalar []uint64) {
-	precomputeOnce.Do(initTable)
-
 	wvalue := (scalar[0] << 1) & 0xff
 	sel, sign := boothW7(uint(wvalue))
 	sm2p256SelectBase(p.xyz[0:8], p256Precomputed[0][0:], sel)
@@ -809,7 +811,7 @@ func Uint64ToAffine(in []uint64) (x, y *big.Int) {
 }
 
 //precompute public key table
-func (curve p256Curve) InitPubKeyTable(x,y *big.Int) (Precomputed *[37][64*8]uint64) {
+func (curve p256Curve) InitPubKeyTable(x, y *big.Int) (Precomputed *[37][64 * 8]uint64) {
 	Precomputed = new([37][64 * 8]uint64)
 
 	var r p256Point
@@ -822,9 +824,9 @@ func (curve p256Curve) InitPubKeyTable(x,y *big.Int) (Precomputed *[37][64*8]uin
 	r.xyz[10] = 0x0000000000000000
 	r.xyz[11] = 0x0000000100000000
 	basePoint := []uint64{
-		r.xyz[0], r.xyz[1],r.xyz[2],r.xyz[3],
-		r.xyz[4],r.xyz[5],r.xyz[6],r.xyz[7],
-		r.xyz[8],r.xyz[9],r.xyz[10],r.xyz[11],
+		r.xyz[0], r.xyz[1], r.xyz[2], r.xyz[3],
+		r.xyz[4], r.xyz[5], r.xyz[6], r.xyz[7],
+		r.xyz[8], r.xyz[9], r.xyz[10], r.xyz[11],
 	}
 	t1 := make([]uint64, 12)
 	t2 := make([]uint64, 12)
@@ -864,7 +866,7 @@ func (curve p256Curve) InitPubKeyTable(x,y *big.Int) (Precomputed *[37][64*8]uin
 }
 
 //fast sm2p256Mult with public key table
-func (p *p256Point) p256PreMult(Precomputed *[37][64*8]uint64, scalar []uint64) {
+func (p *p256Point) p256PreMult(Precomputed *[37][64 * 8]uint64, scalar []uint64) {
 	wvalue := (scalar[0] << 1) & 0xff
 	sel, sign := boothW7(uint(wvalue))
 	sm2p256SelectBase(p.xyz[0:8], Precomputed[0][0:], sel)
@@ -907,13 +909,13 @@ func (p *p256Point) p256PreMult(Precomputed *[37][64*8]uint64, scalar []uint64) 
 }
 
 //fast scalarmult with public key table
-func (curve p256Curve) PreScalarMult(Precomputed *[37][64*8]uint64, scalar []byte) (x,y *big.Int) {
+func (curve p256Curve) PreScalarMult(Precomputed *[37][64 * 8]uint64, scalar []byte) (x, y *big.Int) {
 	scalarReversed := make([]uint64, 4)
 	p256GetScalar(scalarReversed, scalar)
 
 	r := new(p256Point)
-	r.p256PreMult(Precomputed,scalarReversed)
-	x,y = r.p256PointToAffine()
+	r.p256PreMult(Precomputed, scalarReversed)
+	x, y = r.p256PointToAffine()
 	return
 }