add preload once for table prepare
adjust some big int pointer to avoid meaning less creation
adding const to avoid create when ZById

Signed-off-by: SamYuan1990 <yy19902439@126.com>