@SamYuan1990 2021-02-01
before
```
BenchmarkSign-2               	   39996	     31654 ns/op	    2155 B/op	      38 allocs/op
BenchmarkEcdsaSign-2          	   36830	     32995 ns/op	    2636 B/op	      32 allocs/op
BenchmarkVerify-2             	   23694	     48765 ns/op	    2448 B/op	      48 allocs/op
BenchmarkEcdsaVerify-2        	   11841	    103615 ns/op	     824 B/op	      16 allocs/op
BenchmarkSignWithDigest-2     	   49904	     23366 ns/op	    1782 B/op	      27 allocs/op
BenchmarkVerifyWithDigest-2   	   31393	     43193 ns/op	    1680 B/op	      23 allocs/op
BenchmarkSignWithASN1-2       	   34216	     32999 ns/op	    2828 B/op	      62 allocs/op
BenchmarkVerifyWithASN1-2     	   24733	     49708 ns/op	    2432 B/op	      48 allocs/op
```
after
```
BenchmarkSign-2               	   45579	     26405 ns/op	    1970 B/op	      32 allocs/op
BenchmarkEcdsaSign-2          	   37306	     31758 ns/op	    2636 B/op	      32 allocs/op
BenchmarkVerify-2             	   30076	     39516 ns/op	    2232 B/op	      43 allocs/op
BenchmarkEcdsaVerify-2        	   12909	     92833 ns/op	     824 B/op	      16 allocs/op
BenchmarkSignWithDigest-2     	   52122	     23128 ns/op	    1694 B/op	      25 allocs/op
BenchmarkVerifyWithDigest-2   	   34992	     34330 ns/op	    1632 B/op	      22 allocs/op
BenchmarkSignWithASN1-2       	   42319	     28492 ns/op	    2654 B/op	      56 allocs/op
BenchmarkVerifyWithASN1-2     	   30667	     39228 ns/op	    2264 B/op	      44 allocs/op
```

reduced cpu ops and memory usage for BenchmarkSign, BenchmarkVerify ,BenchmarkSignWithDigest, BenchmarkVerifyWithDigest, BenchmarkSignWithASN1, BenchmarkVerifyWithASN1